# openapi.yaml
openapi: 3.0.3
info:
  title: Varius API
  version: 0.1.0
  description: "OpenAPI specification for the vx-api-server (summary of implemented endpoints)"
servers:
  - url: https://api.varius.technology
    description: Production server
paths:
  /auth:
    get:
      summary: Redirect to GitHub OAuth
      parameters:
        - name: redirect_url
          in: query
          schema:
            type: string
          description: Optional URL to redirect the user to after auth
      responses:
        '302':
          description: Redirect to GitHub OAuth page
  /auth/github/callback:
    get:
      summary: OAuth callback endpoint
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirect after successful login
        '400':
          description: Missing authorization code
        '500':
          description: Internal server error during authentication
  /auth/login:
    post:
      summary: API login (credential-based)
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Login success
  /auth/status:
    post:
      summary: Check auth status (legacy)
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Status
...
  /auth/me:
    get:
      summary: Get authenticated user from signed cookie
      responses:
        '200':
          description: Returns session user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /redirect:
    get:
      summary: Redirect helper
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
        - name: user
          in: query
          schema:
            type: string
      responses:
        '302':
          description: Redirect to provided URL
        '400':
          description: Invalid url parameter
  /users/{id}:
    get:
      summary: Get user by numeric ID (GitHub ID)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFoundResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        msg:
          type: object
    User:
      type: object
      properties:
        id:
          type: integer
        login:
          type: string
        name:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
    UserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
    UserFoundResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
...